requirements.txt
Django==5.1.3
mysql-connector-python==9.2.0
Pillow==11.1.0
python-decouple